Moved the Message and Thread components.
