+++ src/app/components/messaging/thread/thread.component.ts
