+++ src/app/routes/messages/messages.component.ts
@@ -45,8 +45,8 @@ import { SWManager } from "@app/services/sWManager.service";
 import { ApiClientService } from "@app/services/apiClient.service";
 import { Loader } from "@common/loader/loader.component";
 import { ItemDeleteForm } from "@forms/itemDeleteForm/itemDeleteForm.component";
-import { AppSingleMessage } from "@app/components/common/message/message.component";
-import { AppSingleThread } from "@app/components/common/thread/thread.component";
+import { AppSingleMessage } from "@app/components/messaging/message/message.component";
+import { AppSingleThread } from "@app/components/messaging/thread/thread.component";
 
 interface MessagesResponse {
   success: boolean;
